Added No Help Hours Today message
When it is saturday or sunday we will never have help hours (because we
like sleeping in, or not sleeping at all whilst on a coding binge). If
there are no help hours for a day, then instead of displaying a table
that has no rows, the page now displays a nice message and gives a link
to the contact area.