This is the working live version of the server.
The .htaccess files are in the gitignore which means that
YOU SHOULD NOT FUCK WITH THEM. Seriously. Don't mess with the .htaccess
it's the thing that's running the show and you shouldn't mess with it if
you don't know what you're doing. Thanks!

There are a few things to note. First off, the router. If you're developing
for the site locally. You're going to change it. You should remember to go
ahead and change things back to the live configuration. And dont delete the
live configuration, just comment it out. YES we're under version control, so
technically you could go grab the changes from a prior commit. But hey, not
everyone is so clever. So leave that shit commented out so you dont forget
to change it.

ALSO.

This is a working live version, but the url's on the hours page are fucked up
that's my fault for not using the BASE DIR in my urls. So. Sorry about that.
I'll get on that.

If you've read all of this. Congratulations.